moved msg forwarding to plugins manager, moved its test cases
